MagaLabs_LogPrint_Web/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
autopep8==2.3.2
BareNecessities==0.2.8
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
decorators==2.0.7
distro==1.9.0
executing==2.2.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
frozenlist==1.8.0
funcs==0.11.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.61.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipython==8.37.0
itsdangerous==2.2.0
javascript==1!1.2.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.9
langchain-core==1.2.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
Mail==2.1.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.0
named==1.4.2
narwhals==2.16.0
notifications==0.3.2
numpy==2.3.4
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pefile==2023.2.7
pg8000==1.31.5
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pylance==0.38.3
pyparsing==3.3.2
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
pywinrm==0.5.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.7
scramp==1.4.8
services==0.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solus==1.2.2
SQLAlchemy==2.0.44
sspilib==0.5.0
stack-data==0.6.3
streamlit==1.54.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-aliases==1.10.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wraps==0.15.0
WTForms==3.2.1
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
prometheus-client==0.20.0
python-json-logger==2.0.7
